+++ src/com/company/banking/Main.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.banking;
 
 // Author: Przemyslaw Reducha, Dominik Rowinski
