+++ src/com/company/banking/models/Client.java
@@ -1,4 +1,4 @@
-package com.company;
+package com.company.banking.models;
 
 // exceptions
